pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beak"
version = "0.1.0"
description = "A toolkit for evolutionary sequence analysis and remote job management"
authors = [{name = "Micah Olivas", email = "micah5051olivas@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Operating System :: OS Independent"
]
dependencies = [
    "numpy",
    "scipy",
    "biopython",
    "gemmi",
    "pandas",
    "fabric>=3.0",
    "paramiko>=3.0",
    "holoviews",
    "matplotlib",
    "seqlogo",
    "logomaker",
    "panel",
    "seaborn",
    "tqdm",
    "ipympl>=0.9.0",
    "cryptography",
    "click>=8.0",
    "rich>=13.0",
    "tomli>=1.0; python_version < '3.11'",
]

[project.scripts]
beak = "beak.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "ipython",
    "jupyter",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
beak = [
    "remote/docker/*",
    "datasets/data/*"
]